[hotfix] [k8s] Make FlinkPodBuilder a internal static class.
