+++ flink-kubernetes/src/main/java/org/apache/flink/kubernetes/kubeclient/decorators/FlinkConfMountDecorator.java
@@ -22,7 +22,6 @@ import org.apache.flink.annotation.VisibleForTesting;
 import org.apache.flink.client.cli.CliFrontend;
 import org.apache.flink.configuration.Configuration;
 import org.apache.flink.kubernetes.kubeclient.FlinkPod;
-import org.apache.flink.kubernetes.kubeclient.FlinkPodBuilder;
 import org.apache.flink.kubernetes.kubeclient.parameters.AbstractKubernetesParameters;
 
 import org.apache.flink.shaded.guava18.com.google.common.io.Files;
@@ -79,7 +78,7 @@ public class FlinkConfMountDecorator extends AbstractKubernetesStepDecorator {
 			.endVolumeMount()
 			.build();
 
-		return new FlinkPodBuilder(flinkPod)
+		return new FlinkPod.Builder(flinkPod)
 			.withPod(mountedPod)
 			.withMainContainer(mountedMainContainer)
 			.build();
